proxy: make compiler happy

+++ proxy/src/dnsproxy_listener.cpp
@@ -276,7 +276,7 @@ public:
             m_state = state::RD_PAYLOAD;
         }
         if (m_state == state::RD_PAYLOAD) {
-            if (m_data.size() < 2 + m_size) {
+            if (m_data.size() < (size_t)2 + m_size) {
                 return false; // Need more data
             }
             out = ag::uint8_vector(m_data.begin() + 2, m_data.begin() + 2 + m_size);
